add access control and filter for photo

+++ api/services/photo/delete.py
@@ -8,9 +8,13 @@ class DeletePhotoService(Service):
         model = Photo
         fields = ['photo_id']
 
-    def process(self):
+    def process(self, author=None):
         try:
             photo = Photo.objects.get(id=self.cleaned_data['photo_id'])
+            if photo.author != author:
+                raise PermissionError(
+                    "У вас нет прав для удаления этого фото."
+                )
             photo.delete()
             return photo
         except Exception:
